refactor(datetime): consolidate all Go files into datetime.go
Merges format_date_time.go, get_relative_time.go, get_timezone_by_country.go,
get_weekdays.go, and locale_support.go into a single datetime.go file.
Tests remain in datetime_test.go unchanged.
